requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
torch>=2.0.0
torchvision>=0.15.0
easyocr==1.7.1
numpy==1.26.4
faiss-cpu==1.8.0  
xgboost==2.0.3
joblib==1.3.2
PyYAML==6.0.1
Pillow==10.2.0